refactor: Cleaned up files
